quiz-jsonを外部クレート化
