+++ Cargo.toml
@@ -2,4 +2,5 @@
 members = [
   "./backend",
   "./frontend",
+  "./quiz_json",
 ]
