+++ backend/Cargo.toml
@@ -14,3 +14,5 @@ actix-web-actors = "3"
 build_timestamp = "0.1.0"
 serde = "1.0.136"
 serde_json = "1.0.78"
+
+quiz_json = {path = "../quiz_json"}
