+++ backend/src/quiz/mod.rs
@@ -4,5 +4,5 @@ mod quiz_json_reader;
 
 #[get("/quiz")]
 pub async fn quiz() -> impl Responder {
-    HttpResponse::Ok().json(quiz_json_reader::read_file())
+    HttpResponse::Ok().json(quiz_json::read_file("quizzes.json"))
 }
