+++ quiz_json/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "quiz_json"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+getset = "0.1.2"
+serde = "1.0.136"
+serde_json = "1.0.78"
