+++ quizzes.json
@@ -1,0 +1,32 @@
+[
+  {
+    "id": 1,
+    "question": "あああああああああああ",
+    "answer": "ああああああああああああ",
+    "explanatory": "ええええええええええええええ"
+  },
+  {
+    "id": 2,
+    "question": "あああああああああああ",
+    "answer": "ああああああああああああ",
+    "explanatory": "ええええええええええええええ"
+  },
+  {
+    "id": 3,
+    "question": "あああああああああああ",
+    "answer": "ああああああああssああああ",
+    "explanatory": "ええええええええええええええ"
+  },
+  {
+    "id": 4,
+    "question": "あああああああああああ",
+    "answer": "ああああああああああaaaああ",
+    "explanatory": "ええええええええええええええ"
+  },
+  {
+    "id": 5,
+    "question": "あああああああああああ",
+    "answer": "ああああああああああああ",
+    "explanatory": "ええええええええええええええ"
+  }
+]
